Code restructuring.
